requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.3
carla==0.9.15
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.0
gast==0.5.3
google-auth==2.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.12.1
idna==3.7
importlib-metadata==7.0.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mkl_fft==1.3.10
mkl_random==1.2.7
mkl-service==2.4.0
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pooch==1.7.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
scikit-learn==1.5.2
scipy==1.13.1
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
termcolor==2.1.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.14.1
yarl==1.11.0
zipp==3.17.0
